requirements.txt
streamlit
mysql-connector-python
passlib